+++ src/client/components/sidebar.js
@@ -1,7 +1,6 @@
 import { html } from '../utils'
 
 import search from './search'
-import { done } from './icons'
 
 const iconSpan = props => html`
   <span>${props.icon}${props.text}</span>
@@ -17,9 +16,13 @@ const menuItem = ({ href, className, icon, text }) => html`
   </li>
 `
 
+const menuHeading = props => html`
+  <h4>${iconSpan(props)}</h4>
+`
+
 const menuList = props => html`
   <div class="menu-list">
-    <h4>${iconSpan(props.heading)}</h4>
+    ${menuHeading(props.heading)}
     <ul>
     ${
       props.items.map(item => menuItem(item))
@@ -28,35 +31,9 @@ const menuList = props => html`
   </div>
 `
 
-const testItems = [
-  {
-    href: 'foo.com',
-    text: 'blaksjljasdf',
-    icon: done()
-  },
-  {
-    href: 'bar.com',
-    text: 'ajsdlfkjalksdjfa',
-    icon: done()
-  },
-  {
-    href: 'baz.com',
-    text: 'ajsdljfla d',
-    icon: done()
-  }
-]
-
-const menu = {
-  items: testItems,
-  heading: {
-    icon: done(),
-    text: 'Recent Posts'
-  }
-}
-
-export default (props = menu) => html`
+export default (props = {}) => html`
   <nav id="sidebar">
     ${search()}
-    ${menuList(menu)}
+    ${props.map(h => menuList(h))}
   </nav>
 `
