+++ src/client/styles/vars.scss
@@ -23,3 +23,8 @@ $sidebar-border: 1px solid blue;
 $sidebar-height: 100%;
 $sidebar-width: 33%;
 $sidebar-bg: darken($bg-default, 10%);
+
+$sb-parent-pad: 0.5 * $base;
+$sb-icon-margin: 0.25 * $base;
+$sb-icon-w: $base;
+$sb-child-pad: $sb-parent-pad + $sb-icon-w + $sb-icon-margin;
